Moved utils to a separate package

+++ core/src/main/kotlin/gregc/gregchess/chess/stats.kt
@@ -2,6 +2,7 @@ package gregc.gregchess.chess
 
 import gregc.gregchess.*
 import gregc.gregchess.registry.*
+import gregc.gregchess.util.*
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.builtins.serializer
@@ -22,8 +23,8 @@ fun ChessGame.addStats(stats: ByColor<PlayerStatsSink>) {
         }
         else -> {}
     }
-    stats.white.add(ChessStat.MOVES_PLAYED, board.fullmoveCounter.toInt() - 1, if (currentTurn == Color.BLACK) 1 else 0)
-    stats.black.add(ChessStat.MOVES_PLAYED, board.fullmoveCounter.toInt() - 1)
+    stats.white.add(ChessStat.MOVES_PLAYED, board.fullmoveCounter - 1, if (currentTurn == Color.BLACK) 1 else 0)
+    stats.black.add(ChessStat.MOVES_PLAYED, board.fullmoveCounter - 1)
     stats.forEach { it.add(ChessStat.TIME_PLAYED, Duration.between(startTime!!, endTime!!)) }
     callEvent(AddStatsEvent(stats))
     stats.forEach { it.commit() }
